Start splitting Body up into modal and tweet components
